[madi_extract_functions] Adding place_ships_on_map, ships_has_no_more_lives AND  has_lost functions
