+++ madi.py
@@ -1,3 +1,7 @@
+from amir import create_map
+from iulia import read_coordinates
+
+
 def display_game_map(game_map):
     for x in range(65, 70):
         print("  " + chr(x) + " ", end="")
@@ -7,3 +11,22 @@ def display_game_map(game_map):
         for col in range(5):
             print(f" {game_map[row][col]}" + " | ", end="")
         print("\n " + "----+" * 5)
+
+
+def place_ships_on_map(ships):
+    game_map = create_map()
+    for ship in ships:
+        [x_axis, y_axis] = read_coordinates()
+
+        mark_ship_on_map(game_map, ship, x_axis, y_axis)
+        display_game_map(game_map)
+    return game_map
+
+
+def ship_has_no_more_lives(game_map, x_axis, y_axis):
+
+    pass
+
+
+def has_lost(map):
+    pass
